Let exceptions escape to the caller if desired
Currently, if a file ails to attach the user of the API gets no
programatic way of dealing with this situation. The API wrapper catches
the xmlrpc fault, logs the error, then does nothing.

Given the funiton as written has no return code semantics it seems
easier to have the exception escape back to the caller if a parameter is
set rather then change the way the function should be interacted with