Allow for custom file name in confluence
The issue with using the filename as given to the function as the
filename of the confluence object it's uploaded as if that filename
contains a path, i.e. ./some/pic.png,  and it's an image then confluence
breaks when you try to reffernce it

Add the option to speicfy the name for the file in confluence by
altering the file dict passed to attachFile from takign a str : str map
and replace the payload with an optional map that allows various options
for each file to be set.

Existing behjviour is preserved, i.e. if you pass a k : v mapping of
filename : comment then that will work and the filename as it will be on
confluence will be the filename of the local file.

If, however, thevalue is passed as a dict the routine will inspect it
for vafioues values.

  currently known keys are

    comment
    confluence_filename
    mimetype

Adding mimetype because trying to infer it from the file extension is
"OK" but not garunteed to be right *at all* and there is a fairly high
chance the calling routine might already know it